refactor: extract url_parser tests to sibling file
